requirements.txt
# Core Scientific Computing
numpy==2.3.5
pandas==2.3.3
scipy==1.16.3

# Machine Learning
torch==2.9.1
scikit-learn==1.6.1

# Visualization
matplotlib==3.10.8
matplotlib-inline==0.2.1

# Jupyter/IPython
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jupyter_client==8.7.0
jupyter_core==5.9.1

# Utilities
joblib==1.5.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
filelock==3.20.1
fsspec==2025.12.0
PyYAML==6.0.2

# Plotting Dependencies
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pillow==12.0.0
pyparsing==3.2.5

# Symbolic Math
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1

# Development Tools
debugpy==1.8.19
jedi==0.19.2
parso==0.8.5

# Terminal/Display
colorama==0.4.6
prompt_toolkit==3.0.52
wcwidth==0.2.14
Pygments==2.19.2

# Async/Process Management
nest-asyncio==1.6.0
psutil==7.1.3
tornado==6.5.4
pyzmq==27.1.0

# Templating
Jinja2==3.1.6
MarkupSafe==3.0.3

# Other
asttokens==3.0.1
comm==0.2.3
decorator==5.2.1
executing==2.2.1
packaging==25.0
platformdirs==4.5.1
pure_eval==0.2.3
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.15.0
tensorboard==2.20.0
